+++ backend/src/main/java/ba/ppis/tim2/backend/service/RecenzijaServiceImpl.java
@@ -27,7 +27,7 @@ public class RecenzijaServiceImpl implements RecenzijaService{
         return recenzijaRepository.findAll();
     }
     @Override
-    public ResponseEntity dodajRecenzijuZaKorisnika(int id, Recenzija recenzija){
+    public ResponseEntity dodajRecenzijuZaKorisnika(Long id, Recenzija recenzija){
         var korisnik = korisnikRepository.findById(id).orElse(null);
         if(korisnik != null) {
             recenzija.setKorisnik(korisnik);
